Actualización de selector de eventos a casos. :wrench: En gráficos.
